+++ flutter_inappwebview/example/integration_test/in_app_webview/main.dart
@@ -89,6 +89,7 @@ part 'set_custom_useragent.dart';
 part 'set_get_settings.dart';
 part 'container_isolation.dart';
 part 'per_webview_proxy.dart';
+part 'container_proxy_change.dart';
 part 'set_web_contents_debugging_enabled.dart';
 part 'should_intercept_request.dart';
 part 'should_override_url_loading.dart';
@@ -205,5 +206,6 @@ void main() {
     containerIsolation();
     perWebViewProxy();
     perWebViewProxyIsolation();
+    containerProxyChange();
   }, skip: shouldSkip);
 }